+++ components/tools/exemples.php
@@ -6,13 +6,17 @@ return array(
         'title' => "Nom de l'outil",
         'description' => "Ceci est la description de l'outil",
         'link' => '#',
-        'link_text' => 'Lien'
+        'link_text' => 'Lien',
+        'color' => '#a2b93b',
+        'icon' => get_template_directory_uri() . '/components/tools/exemple-icon.svg'
       ),
       array(
         'title' => "Nom d'un autre outil",
         'description' => "Ceci est la description de l'autre outil",
         'link' => '#',
-        'link_text' => 'Lien'
+        'link_text' => 'Lien',
+        'color' => '#e16e38',
+        'icon' => get_template_directory_uri() . '/components/tools/exemple-icon.svg'
       ),
     )
   )