+++ components/tools/tools.php
@@ -9,21 +9,37 @@
 
       foreach ($data->items as $item) :
 
-        $item = (object) $item;
+        // Tableau d'objets Outil
+        if (gettype($item) === 'object' && get_class($item) === 'WP_Post') :
 
-        $link = get_permalink($item);
-        $name = $item->post_title;
+          $item->link = get_permalink($item);
+          $item->title = $item->post_title;
 
-        $fields = (object) get_fields($item->ID);
-        // affiche en gras le nom et l'abbréviation
-        $fields->short_description = str_replace($name, '<strong>' . $name . '</strong>', $fields->short_description);
-        $fields->short_description = str_replace($fields->abbr, '<strong>' . $fields->abbr . '</strong>', $fields->short_description);
+          $fields = (object) get_fields($item->ID);
+          $item->description = isset($fields->short_description) ? $fields->short_description : '';
+          // affiche en gras l'abbréviation
+          if (isset($fields->abbr)) $item->description = str_replace($fields->abbr, '<strong>' . $fields->abbr . '</strong>', $item->description);
+          $item->color = isset($fields->color) ? $fields->color : '#666';
+          $item->icon = isset($fields->icon) ? $fields->icon : false;
+          $item->link_text = isset($fields->link_text) ? $fields->link_text : __('Accéder à cet outil', 'telabotanica');
+
+        // Tableau simple
+        elseif (gettype($item) === 'array') :
+
+          $item = (object) $item;
+
+        endif;
+
+        // affiche en gras le nom
+        $item->description = str_replace($item->title, '<strong>' . $item->title . '</strong>', $item->description);
 
         echo '<li class="component-tools-item">';
-          echo '<a href="' . $link . '" class="component-tools-item-icon" style="color: ' . $fields->color . '"><img src="' . $fields->icon . '" alt="Icône de ' . $name . '" class="style-svg" /></a>';
-          echo '<h4 class="component-tools-item-title"><a href="' . $link . '">' . $name . '</a></h4>';
-          echo '<div class="component-tools-item-description">' . $fields->short_description . '</div>';
-          echo '<div class="component-tools-item-link"><a href="' . $link . '" style="color: ' . $fields->color . '">' . $fields->link_text . ' &rsaquo;</a></div>';
+          echo '<a href="' . $item->link . '" class="component-tools-item-icon" style="color: ' . $item->color . '">';
+          if ($item->icon) echo '<img src="' . $item->icon . '" alt="' . sprintf( __('Icône de %s', 'telabotanica'), $item->title ) . '" class="style-svg" />';
+          echo '</a>';
+          echo '<h4 class="component-tools-item-title"><a href="' . $item->link . '">' . $item->title . '</a></h4>';
+          echo '<div class="component-tools-item-description">' . $item->description . '</div>';
+          echo '<div class="component-tools-item-link"><a href="' . $item->link . '" style="color: ' . $item->color . '">' . $item->link_text . ' &rsaquo;</a></div>';
         echo '</li>';
 
       endforeach;